Adds two implementations for adding in soulmate card
Ideally we would use the <template> tag only, but support doesn't
go that far back. As a fallback we create the nodes one at a time.

This is preferable to getting the soulmates API to send us back a
blob of HTML. The point of this exercise is to get rid of the HTML
templates. Doing the generation on the clientside isn't ideal, but
it won't exert much strain performance wise and it allows us to
clean up our soulmates API to provide a single route, which can
return a JSON array of Soulmate objects.