Generate min/max age ranges with suitable defaults
Currently the defaults are 25 and 45 for minimum and maximum age
respectively. These have been kept along with the current age
ranges of 18 - 99. A modest extrapolation of life expectancy
(see source: http://data.worldbank.org/indicator/SP.DYN.LE00.IN)
and an equally modest expectation that these components will last
for ~50 years leads me to suggest that we increase the limit to
139. This companionship-delimiting can be saved for another PR.