Remove libraries/scripts for generating/reading  config
The original design necessitated a config file that mapped
local directories to DFP formats and styles. However, once
a convention for naming styles was detected and enforced,
maintaining a config became redundant as the style and
format names in DFP can be entirely deduced from the
directory structure once the templates have been built
locally.