Create Guardian Masterclass/Live native ad template
This introduces a new ad type that automates the process of creating an
advert for use with masterclasses and live events.

Information is pulled in from a membership API endpoint and this is used
to auto-generate the advert content based on the URL.

At present this points to the live events end point only.

Co-authored-by: Fares Marek Basmadji <fares.basmadji@guardian.co.uk>