Remove redundancy in the code base
The masterclass and live events automated adverts were effectively using
the `[%Tone%]` to determine styling but had instead had this hard coded
into each advert.
This moves to using the `[%Tone%]` variable directly, reducing redundancy
in the code base and making the adverts easier to implement in Google Ad
Manager.